# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/cart/add":{"post":{"summary":"Add a product to the cart","description":"Add a product to the cart by specifying the product code and quantity.","operationId":"add_to_cart_cart_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cart/list":{"get":{"summary":"List all items in the cart with pricing details","description":"Retrieve all products currently in the cart, along with their quantities, original prices, and discounted prices (if applicable).","operationId":"list_cart_cart_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCartResponse"}}}}}}},"/cart/clear":{"post":{"summary":"Clear the cart","description":"Remove all products from the cart.","operationId":"clear_cart_cart_clear_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartResponse"}}}}}}}},"components":{"schemas":{"CartItem":{"properties":{"item":{"type":"string","title":"Item"},"quantity":{"type":"integer","title":"Quantity"},"original_price":{"type":"number","title":"Original Price"},"discounted_price":{"type":"number","title":"Discounted Price"}},"type":"object","required":["item","quantity","original_price","discounted_price"],"title":"CartItem","description":"Represents a single item in the cart with pricing details."},"CartResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"CartResponse","description":"Generic response model for actions like adding or clearing the cart."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ListCartResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/CartItem"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"ListCartResponse","description":"Represents the response for the `/cart/list` endpoint."},"ProductInput":{"properties":{"code":{"type":"string","title":"Code"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["code","quantity"],"title":"ProductInput"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}